requirements.txt
# ============================================================================
# Agent Server - Requirements
# ============================================================================
# Python 3.12+
#
# 安裝: pip install -r requirements.txt
# ============================================================================

# ----------------------------------------------------------------------------
# Core Framework & Async
# ----------------------------------------------------------------------------
asyncio-atexit>=1.0.1
aiohttp>=3.13.0
python-socketio>=5.16.0

# ----------------------------------------------------------------------------
# LangChain & LangGraph
# ----------------------------------------------------------------------------
langchain>=1.2.0
langchain-core>=1.2.0
langchain-openai>=1.1.0
langchain-community>=0.4.0
langchain-text-splitters>=1.1.0
langgraph>=1.1.0
langgraph-checkpoint>=4.0.0
langgraph-checkpoint-postgres>=3.0.0
langgraph-prebuilt>=1.0.0

# ----------------------------------------------------------------------------
# OpenAI & LLM
# ----------------------------------------------------------------------------
openai>=2.29.0

# ----------------------------------------------------------------------------
# Database - PostgreSQL
# ----------------------------------------------------------------------------
asyncpg>=0.31.0
psycopg[binary]>=3.3.0
psycopg-binary>=3.3.0
psycopg-pool>=3.3.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# ----------------------------------------------------------------------------
# Database - Vector Store (Qdrant)
# ----------------------------------------------------------------------------
qdrant-client>=1.17.0

# ----------------------------------------------------------------------------
# Data Validation & Models
# ----------------------------------------------------------------------------
pydantic>=2.12.0
pydantic-settings>=2.13.0

# ----------------------------------------------------------------------------
# Security & Crypto
# ----------------------------------------------------------------------------
cryptography>=46.0.0
python-dotenv>=1.2.0

# ----------------------------------------------------------------------------
# HTTP & API
# ----------------------------------------------------------------------------
httpx>=0.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# ----------------------------------------------------------------------------
# Tools & Utilities
# ----------------------------------------------------------------------------
duckduckgo-search>=6.3.0
python-multipart>=0.0.20

# ----------------------------------------------------------------------------
# Testing
# ----------------------------------------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0

# ----------------------------------------------------------------------------
# Code Quality (Optional - Development)
# ----------------------------------------------------------------------------
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.11.0

# ----------------------------------------------------------------------------
# Logging & Monitoring (Optional)
# ----------------------------------------------------------------------------
python-json-logger>=3.2.0
